+++ libs/react/src/components/molecules/surveyQuestionContainer/surveyQuestionContainer.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
-import { SurveyInput } from '../index';
-import { cloneDeep, findIndex, forEach, forOwn } from 'lodash';
+import { BaseButton, ErrorFallback, SurveyInput } from '@coldpbc/components';
+import { cloneDeep, findIndex, forEach, forOwn, isArray, isEqual } from 'lodash';
 import { IButtonProps, SurveyActiveKeyType, SurveyAdditionalContext, SurveyPayloadType, SurveySectionType } from '@coldpbc/interfaces';
-import { BaseButton, Spinner } from '../../atoms';
 import { ButtonTypes, GlobalSizes } from '@coldpbc/enums';
 import { CSSTransition, SwitchTransition } from 'react-transition-group';
 import { getSectionIndex, isComponentTypeValid, isKeyValueFollowUp } from '@coldpbc/lib';
 import { axiosFetcher } from '@coldpbc/fetchers';
 import { withErrorBoundary } from 'react-error-boundary';
-import { ErrorFallback } from '../../application';
 import { useAuth0Wrapper } from '@coldpbc/hooks';
 
 export interface SurveyQuestionContainerProps {
@@ -365,9 +363,10 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
         }
       }
     } else {
-      if (additionalContextQuestion && sections[activeSectionKey].additional_context &&
-        (sections[activeSectionKey].additional_context?.value === undefined
-          || sections[activeSectionKey].additional_context?.value === null)
+      if (
+        additionalContextQuestion &&
+        sections[activeSectionKey].additional_context &&
+        (sections[activeSectionKey].additional_context?.value === undefined || sections[activeSectionKey].additional_context?.value === null)
       ) {
         buttonProps.disabled = true;
       }
@@ -602,7 +601,11 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
   const ifAdditionalContextConditionMet = (value: any, additionalContext: SurveyAdditionalContext) => {
     switch (additionalContext.operator) {
       case '==':
-        return value === additionalContext.comparison;
+        // make comparison for arrays if both the value and comparison are arrays
+        if (isArray(value) && isArray(additionalContext.comparison)) {
+          return isEqual(value, additionalContext.comparison);
+        }
+        return isEqual(value, additionalContext.comparison);
       case '!=':
         return value !== additionalContext.comparison;
       case '>':
@@ -613,6 +616,32 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
         return value >= additionalContext.comparison;
       case '<=':
         return value <= additionalContext.comparison;
+      case 'in':
+        // check if the value is in the comparison array
+        if (isArray(additionalContext.comparison)) {
+          if (!isArray(value)) {
+            return additionalContext.comparison.includes(value);
+          } else {
+            // check if any values in the value array are in the comparison array
+            return value.some((val: any) => {
+              return additionalContext.comparison.includes(val);
+            });
+          }
+        }
+        return false;
+      case 'has':
+        // check if the comparison value is in the value array
+        if (isArray(value)) {
+          if (!isArray(additionalContext.comparison)) {
+            return value.includes(additionalContext.comparison);
+          } else {
+            // check if any of the values in the comparison array are in the value array
+            return additionalContext.comparison.some((val: any) => {
+              return value.includes(val);
+            });
+          }
+        }
+        return false;
       default:
         return false;
     }